postgres7: fix system warning in MetaColumns() with schema
When calling MetaColumns with schema.table, the driver triggers a
'Warning: sprintf(): Too few arguments' because the metaColumnsSQL1
statement contains one more parameter in 7 compared to the 6.4 version.

Resolved by adding a new protected _generateMetaColumnsSQL() method to
handle the preparation of the appropriate SQL statement, and overriding
it in the postgres7 driver.

http://phplens.com/lens/lensforum/msgs.php?id=19481