Add support for pagination with complex queries
There are cases where ADODB is not be able to figure out the proper
count(*) queries for PageExecute() on its own.

To work around this problem, ADOdb now handles a new keyword
"_ADODB_COUNT" that can be wrapped around the parts of the SQL query
that should be replaced with "COUNT(*)"; the code will remove everything
between and including the _ADODB_COUNT keywords and replace it with
COUNT(*). The _ADODB_COUNT tags are of course removed when COUNT(*) is
not needed.

Example usage:
    SELECT
        _ADODB_COUNT
        col1,
        col2,
        col3,
        ( select col4 from table2 where col1=col2 )
        col4,
        _ADODB_COUNT
    FROM table1
    LEFT JOIN ( select col5 from table3 where col3=col4 ) AS tmp;

Original 2012 bug report: http://phplens.com/lens/lensforum/msgs.php?id=19058
The fix has been in use since that time in hundreds of thousands of
production installations.

Fixes #88

Signed-off-by: Damien Regad <dregad@mantisbt.org>