MSSQL: fix fail on insert when statement contains ';'
In mssql a tick mark (') in a value string is escaped via a second one.
Therefore the number of tick marks before the semicolon must be odd
(since the string starts with a tick mark).

Semicolons should be allowed after an odd number of tick marks.
There must be a character between the odd tick mark groups.

Fixes #97, see also #96