sqlite3: Fix result-less SQL statements executed twice
The sqlite3 driver relies on SQLite3::query to execute the statement,
which returns an empty SQLite3Result while the ADOConnection::_Execute()
method expects true for handling of result-less queries. This causes the
query to be executed a second time when ADOdb initializes the recordset
for what it wrongly assumes to be a SELECT statement.

This also addresses the same behavior with the sqlite driver, but the
fix was not tested as I don't have a sqlite 2 setup available.

Fixes #99