(Same change as previous commit, but for different sqlite driver)
Sqlite has automatically created indexes which can be safely assumed to be unaccounted for by calling code, such as AXMLS, as well as different to normal indices in terms of what may be done with them. Without this fix, a number of failures happen starting with the line

"$cols = explode(")",$cols[1]);"
later in the same function, as well as in AXMLS at least.