oci8po: prevent replacement of '?' within strings
When processing an SQL statement containing a '?' character within a
string, ADOdb wrongly assumed it is for a bind variable.  The number of
variables then no longer matches the number passed, and crashes the
_query() function.

We now identify whether a '?' is within a string and don't use it as a
bind variable if so.

Fixes #132

Signed-off-by: Damien Regad <dregad@mantisbt.org>