Fix incorrect handling of input array in Execute()
When executing a query that does not expect any parameters with a
non-empty input array, the parameters are added at the end of the
query.

Example:
- $db->Execute('select * from table where id=', 1);
  expected: failure (extra param, invalid query)
  actual:   execution of 'select * from table where id=1'
- $db->Execute('select * from table where id=1', 1);
  expected: failure (extra param)
  actual:   execution of 'select * from table where id=11'

This has been confirmed to occur using mysqli driver, but could possibly
affect others drivers with _bindInputArray = false as well.

Execute() now issues an error when the number of provided input
parameters does not match what the query expects.

Thanks to peterdd for the initial research and proposed fix.

Fixes #146