sqlite3 driver: use -1 for _numOfRows
This relies on the superclass's built-in default of $this->_numOfRows
being set to -1. That causes the sqlite3 driver to rely on
ADODB_COUNTRECS.

Before this patch, we were setting _numOfRows to positive 1, which meant
that any query that checked how many rows were available in a queryset
would see literally the number 1, which was not accurate for queries that
returned 0 rows or queries that returned more rows.

In the case of sqlite3, there appears to be no efficient way to ask the
PHP sqlite3 driver how many rows come back from a query, so relying on
ADODB_COUNTRECS is the best we can do.