Upload script: fix bug with -d / -f options
The behavior was reverted, i.e. -d would disable documentation upload
(instead of files), and vice-versa.

Move check for version number check for files updload within the if
block to avoid an error when uploading the documentation when the
release tarballs are not present (e.g. from the repository's root).