ADOConnection::Version() now handles SemVer
In the past, ADOdb versions were numbered X.YY, and hot fixes were
identified by an optional letter suffix, e.g. 5.17, 5.18a. Starting with
5.20, we switched to Semantic Versioning.

The regex in ADOConnection::Version() was not adjusted to reflect this,
and the function incorrectly returns '5.20' instead of '5.20.0'.

The code now correctly handles SemVer version numbers, and also allows
the following pre-production suffixes: dev, alpha, beta, rc; the last 3
must be followed by an additional version number (e.g. 5.21.0-beta.1).

In case the ADOdb version string (defined in $ADODB_vers global) does
not match the regex, the function outputs a warning and falls back to
returning a substring starting with 1st character (excluding leading 'v'
if present), until the first whitespace or the end of the string).

Fixes #164