mysqli: method failed if $associative set true
Passing the associative flag attempted to control both how the data was
retrieved from the database and how it was returned from the method.
This caused inconsistencies due to the number of different combinations
available. Flag now only controls presentation of data, retrieval is
always done in same way.

Fixes #181