mssqlnative: Query not returning id
The existing methodology, which has been in place since version 5.19,
creates a substantial overhead in executing queries, both in the regexp
parsing of queries to determine if the query is an insert statement, as
well as appending an attempt to return an insert_id, even when neither
available or needed.

This patch re-implements the _query method from Version 5.18, which is
faster than the current version, but also adds an optimized retrieval of
the last insert id via the _insertid method.

Fixes #185