mssqlnative: avoid repeated sqlsrv_field_metadata() calls
The mssqlnative driver was poorly structured, _fetchField() repeatedly
called sqlsrv_field_metadata() to retrieve metadata from the database.

The driver has been restructured to resemble the method used by the oci8
driver, in that the metadata function is called once when the recordset
is initialized. Subsequent requests for information are read from a
class variable that holds all the table metadata.

Fixes #70