addColumnSql now supports ENUM data types
addColumnSql (and methods that call it, such as createTableSql), now
correctly parse the ENUM data type and associated VALUE lists, if the
underlying database supports it. actively supported databases known to
support it are mysql, postgres, oracle, db2.

The statement will fail if the database does not support it (e.g. mssql,
sqlite)

Fixes #26