Enhancement to getInsertSQL() value defaulting
This enhancement provides an additional method of defaulting values in
getInsertSql(): ADODB_FORCE_NULL_AND_ZERO.

It allows getInsertSql to default columns that have a number type to
zero instead of null when a field is not provided to the insert
statement.

This feature is especially useful for financial applications, when math
may involve the addition of columns, some of which may be null.

Fixes #214