Quarter of year not returned correctly
The ADOdb date/time library includes support for the time format 'Q'
which returns the quarter of the year that the provided date falls into.

This format is not supported by the standard PHP library, which means
that if the provided date falls into the 32-bit date range where ADOdb
hands off data calculations to the standard gmdate() for performance
reasons, then the value is not calculated.

The solution would be that, if a supplied format contained the 'Q'
character, it would bypass the normal 32 bit test and be handled by the
adodb date library in adodb_date()

In addition, the quarter function does not return the correct value, it
returns 3rds of a year instead of Quarters.

Fixes #222