Add week of year support in date library
The ADOdb date/time library did not include support for the time format
'W' which returns the week of the year (1-52) that the provided date
falls into.

This format is supported by the standard PHP library, which means that
if the provided date falls outside the 32-bit date range where adodb
hands off data calculations to the standard gmdate for performance
reasons, then the value is not calculated.

'W' is one of the standard identifiers used by sqlDate(), and adodb_date
is used by sqlite to perform date calculations.

Fixes #223