Fix fatal error when connecting with missing extension
When connecting to a database using mysqli driver and the corresponding
PHP extension is extension is not loaded, ADOdb dies with exit code 255
signaling a fatal error. This is due to ADODB_mysqli::ErrorNo() calling
mysqli_connect_errno() with the '@' operator, suppressing the 'Call to
undefined function' fatal error.

When not using adodb-exceptions, setting `$db->debug=true;` causes ADOdb
to print message 'Missing extension for mysql', but this is not the case
when exceptions are used.

To fix the problem, we only call the ErrorNo() method when the result of
_connect() is set, and otherwise pass arbitrary error number 0 with the
missing extension message to the exception handler.

Fixes #254