Abort with error if ADOdb Extension is detected
When the old ADOdb extension is installed and a more recent version of
ADOdb is running, a PHP warning 'undefined constant
ADODB_ASSOC_CASE_NATIVE' is thrown.

The PHP Extension is obsolete and unsupported, so we shouldn't even try
to run when it is loaded.

This commit will cause ADOdb to exit with an error (or throw an
exception if adodb-exceptions.inc.php has been included) if the
Extension is detected.

Fixes #269