mssql: fix drop/alter column with existing default constraint
With MSSQL it is not possible to drop or alter a column with an
existing constraint. The constraint has to be removed before the
operation takes place.

In 'datadict-mssqlnative.inc.php' AlterColumnSQL is commented out and
DropColumnSQL doesn't care.

This tries to fix the problem with the smallest possible impact:

- Fix DropColumnSQL(), to allow the drop even with a constraint on that
  given column. We drop, so any default doesn't matter.
- Fix AlterColumnSql(), to allow changes if (and only if !) either a
  'new' default is given for an existing default, or there is no
  existing one at all. So something like
  'ALTER TABLE t ALTER COLUMN c INT NOT NULL'
  with an existing constraint on c will still fail since it can't be
  determined if keeping or removing the constraint is implied here.

Fixes #290 via #297

Changes to original commit:
- split long lines, whitespace
- Added commit message text from issue #290's description

Signed-off-by: Damien Regad <dregad@mantisbt.org>