Added support for connecting using windows authentication
To connect to the SQL Server using windows authentication the username and password arguments have to be omitted. So i just check if these are null or have a length of 0. If both the username and password are not set I do not add them to the connection info. This way the native SQL driver connects using windows authentication.