Use include_once instead of include (#361)
Most of the includes in adodb.inc.php are done with include_once but
in a couple locations its still done with include.

This is the case for adodb-lib.inc.php which is conditionally included
based on a global var. I ran into problems with running adodb in debug
mode while running phpunit tests. Phpunit will reset the global vars
and thus break the include mechanism. This is a problem of phpunit
and can be worked around. However it would make adodb more robust
to just use include_once everywhere. It will remove the dependency on
the $ADODB_INCLUDED_LIB global var.

I also replaced one include statement in adodb-time.inc.php, in various 
drivers and other files too, so include_once is used everywhere.

Signed-off-by: Damien Regad <dregad@mantisbt.org>