pgsql: new ServerInfo() param to avoid unnecessary SELECT
The PostgreSQL server version number can be retrieved through
pg_parameter_status().

A new optional parameter was added to allow caller to only retrieve the
version number without executing an extra SQL statement.

To maintain backwards compatibility and keep returning a detailed
version string in the 'description' key of the version array as the code
used to do, the parameter defaults to true.