PHP-X-version compat: Don't bother with magic quotes when not available
Magic quotes were removed from PHP in v5.4.0.

The return value of `get_magic_quotes_runtime()` will be 0 if `magic_quotes_runtime` is off, 1 otherwise.
And as of PHP 5.4.0, it will always returns `false`.

This minor change prevent trying to set the ini value when the ini directive is not available and cleans up some old/unused/commented out code.

Refs:
* http://php.net/manual/en/info.configuration.php#ini.magic-quotes-runtime
* http://php.net/manual/en/function.get-magic-quotes-runtime.php