Fix LDAP connections using URIs
Commit 53d13c0f64ad97ff87fa3b74fa4a4e896e691c9b (fixing #58 and #63)
introduced a regression in 5.20.0 that broke connections LDAP when using
URIs like ldap:// or ldaps://.

Create a new protected method ADOConnection::parseHostNameAndPort() to
help parse off the port from a hostname while retaining the ldap:// or
ldaps:// scheme, and use it to refactor Connect() and PConnect().

Fixes #340

Signed-off-by: Damien Regad <dregad@mantisbt.org>

The PR's original commits were squashed and the message reworded. The
fix (initially targeted at master) was applied to hotfix/5.20 branch.