mssql: MetaForeignKeys not returning all FK's
MetaForeignKeys returns an associative array with the name of reference
table for key. But when there are 2 foreign keys from the same reference
table, MetaForeignKeys returns only the last foreign key.

A solution is to merge arrays when there are several foreign keys.

Fixes #486

Signed-off-by: Damien Regad <dregad@mantisbt.org>

Original commits from PR squashed and message reworded.