Remove unnecessary srand() calls
srand() should not be called manually, unless one has a better way to
initialise the already-initialised random seed.

The way the current srand() call is made is buggy, and will lead to
duplicate random seeds if called often enough. Because the current call
will lose the timestamp part, and will only use the first 6 chars from
the current microsecond, but the system might not have a high-enough
timer resolution, so the actual pool of random seeds gets even smaller.

Fixes #532

Signed-off-by: Damien Regad <dregad@mantisbt.org>