Fix Active Record field names case
When I have a class extending ADODB_Active_Record, I have only lowercase
field names instead native-case field names. When the value of
ADODB_ASSOC_CASE is 2, the value of $ADODB_ASSOC_CASE is NULL, so we are
going to the case 0 of the switch!

Signed-off-by: Damien Regad <dregad@mantisbt.org>

This is the original code from PR #535, which only addresses part of the
problem. the rest will be fixed in a subsequent commit.

Issue #536