New getMenuGrouped() method, deprecate getMenu3()
The new function is the same as getMenu3(), except that it has an extra
parameter $compareFirstCol to align it with getMenu(). The new param
defaults to true, i.e. defaults are compared against the value, not the
description (opposite of getMenu3()).

getMenu3() is marked as deprecated and just calls the new function.