oci8: create compact trigger/sequence names
This enhancement provides a method of creating an emulated auto-
increment column for a table with a name greater than 25 characters,
built using createTableSql() and specifying an "AUTO" option on the
required column.

Instead of using the table name, a CRC32 value of the tablename is
created and used for the name of the sequence and trigger, i.e. instead
of a trigger called "TRIG_EMPLOYEES" for a table called "EMPLOYEES", it
creates a trigger "TRIG_3129131776", and similarly, a trigger for a
table named "EMPLOYEE_HOURLY_PAY_RATES" becomes "TRIG_3754194288". The
name of the trigger and sequence can be easily obtained after the fact
by finding the CRC32 value of the table name.

The feature is provided with a compatibility flag that must be enabled.
Existing configurations that already use this feature are unchanged. To
activate, set the dictionary class variable
`$dict->useCompactAutoIncrements = true` when building the table, and
the connection class variable
`$db->useCompactAutoIncrements = true` when using the table in ADOdb.

The feature is global for the database and cannot be specified on a
table level basis.

Fixes #565