Cache Field objects with Property instead of static var
Change the static array introduced in fieldTypesArray() (see #367) to a
class variable ($fieldObjectsCache), so the cache now only lasts for the
Recordset's lifetime.

Since the mssqlnative was already using a private $fieldObjects property
for the same purpose, the declaration was removed, and usages of the old
name changed to match the one defined in the parent class.

Fixes #687