Remove subqueries without using regex
Following discussion in #718, using regex to identify and remove
sub-queries in _adodb_getcount() is flawed and can't be fixed. This led
to implementing a hack using _ADODB_COUNT "keyword" to (see #88
41839f5c18abe262ee088f1d4bd92453616d2a4a).

This replaces the regex parsing by a simple loop on the query string,
removing everything between the SELECT and the FROM clause, ignoring
whatever is between parentheses (i.e. subqueries).

The _ADODB_COUNT logic is therefore obsolete

Fixes #715