Optimise _fixblobs()
The call to `pg_fetch_array()` returns `false` in all cases where we wouldn't
need or want to potentially call `_fixblobs()` and return `true`.

A straight up boolean check is all that is needed here.

This occurs in a number of places and all had the same pattern of checks
before the call.  These were moved into `_fixblobs()` and that was
updated to return a boolean should we want to check that.