Refactored `_fixblobs()`
The call to `pg_fetch_array()` returns `false` in all cases where we would not
need or want to potentially `_fixblobs()` and return `true`.

A straight up boolean check is all that is needed here.

Tracking through `_fixblobs()` is called in a number of places (thanks
for the pointer @dregad) and how it is checked for was repeated in every
case.  The checks were moved into `_fixblobs()`.

It was also noted that in every case the setting of `$this->fields` was
identical as well.  `_prepfields()` was added which sets `$this->fields`
and calls `_fixblobs()`.

All instances of `_fixblobs` was replaced with `_prepfields()` and the
`pg_fetch_array()` call deleted.