Refactored ADODB_postgres7::_fixblobs()
The call to pg_fetch_array() returns false in all cases where we would 
not need or want to potentially _fixblobs() and return true. A straight 
up boolean check is all that is needed here, no need for in_array().

Tracking through _fixblobs() is called in a number of places and how its
result is checked was repeated in every case. The checks were moved into
the function, which was updated to return a boolean should we want to
check that.

It was also noted that in every case the setting of $this->fields was 
identical as well so new _prepfields() method was added which sets 
$this->fields and calls _fixblobs().

All instances of _fixblobs() were replaced with _prepfields() and the
pg_fetch_array() call deleted.

Fixes #767 (replacing #760)