Fix affected_rows() returning false
When an update or insert query is executed after a select query,
ADODB_mysqli::$isSelectStatement is not reset causing subsequent calls
to affected_rows() to return false instead of the expected number of
rows.

This was marked as fixed in 5.22.1 [1], but in fact it wasn't so the
Changelog has been updated accordingly.

Fixes #820

[1]: see commit aa88f92f95454fdfda3c4a916f7aac68c74c19a3