Fix PHP Warning when connection is already closed
If errorMsg() is called after closing the MySQL connection (with
mysqli_close()), the method will throw a PHP warning:
Uncaught Error: mysqli object is already closed.

With PHP < 8.0 the warning won't be thrown due to the @ operator, but
the function will return false; with PHP 8.0 or later, this is a fatal
error (exception).

Remove unnecessary @ operator.

Fixes #842