mysqli: store error msg/code after query execution
ADOConnection's errorNo() and errorMsg() methods were returning 0 / ''
when executing an SQL statement with bound parameters, since the
introduction of true prepared statements (in #655).

This is because they are getting error information from the mysqli
connection object, but when using prepared statements, it must be
retrieved from the mysqli_stmt object.

Fixes #872