Remove deprecated $databaseName property
ADOConnection::$databaseName was marked as obsolete in ADOdb 4.66.

Since the assignment that was kept for backwards-compatibility in the
ADOConnection::selectDB() method is causing deprecation warnings with
PHP 8.2, now is the time to get rid of it.

The 2018 rewrite of DB2 driver introduced a private $databaseName
property; this has been removed in favor of ADOConnection::$database.

Fixes #932, #904

Signed-off-by: Damien Regad <dregad@mantisbt.org>