oci8: replace usage of deprecated COL view
Update the $metaColumnsSQL statements in oci8, oci8po, pdo_oci and
odbc_oracle drivers to use USER_TAB_COLUMNS instead of COL. Change was
not applied to the deprecated oracle driver.

Since the contents of the COL.NULLS column (`NULL` / `NOT NULL`) is not
the same as the equivalent USER_TAB_COLUMNS.NULLABLE (`Y` / `N`), we
can get rid of the CASE statement in $metaColumnsSQL2 and simplify the
processing in metaColumns() method, to use a simple string comparison
instead of a strncmp() call to populate ADOFieldObject's not_null
property.

Fixes #1006