Add missing ADOConnection::_*connect() methods
Internal, low-level connection methods are driver-specific, but they
nevertheless need to be declared in the parent ADOConnection class, in
order to define a common signature and avoid potentially polymorphic
calls.

_nconnect() exists, but the _connect() and _pconnect() methods were
missing.

The _connect() method is defined as abstract to force its implementation
in the child Driver classes.

Since they are for internal use, all 3 methods are declared as protected
(in ADOConnection only, not in child classes - for now). Note that this
does not actually break backwards compatibility, as PHP allows child
classes to relax the visibility (i.e. redeclare the methods as public).
At some point we'll need to align this across the library.

Fixes #1013