Fix autoExecute() skipping updates
AutoExecute fetches a row from the table prior to performing an insert
or update. When $forceUpdate is false, only modified columns should be
updated. However, the WHERE clause was not being used when fetching the
row, so the comparison was always made against an arbitrary row, causing
differences to be detected or missed by chance.

This change makes use of the $where parameter when fetching the row from
the table, making the comparison work as intended.

Fixes #1146